+++ UVDI152022/inc/itfe/EItfcLuria.cpp
@@ -2210,7 +2210,7 @@ API_EXPORT BOOL uvEng_Luria_ReqGetActiveAreaQualifier()
  retn : TRUE or FALSE
  note : Note that setting AbsWorkRange will clear the AbsWorkRangeStatus for the given photo head(s).
 */
-API_EXPORT BOOL uvEng_Luria_ReqSetAbsWorkRangeLM(UINT8 ph_no, DOUBLE abs_min, DOUBLE abs_max)
+API_EXPORT BOOL uvEng_Luria_ReqSetAbsWorkRangeLM(UINT8 ph_no, INT32 abs_min, INT32 abs_max) 
 {
 	/* Check if it is the emulated (demo) mode ... */
 	if (GetConfig()->IsRunDemo())	return TRUE;
